+++ src/iteration/edges.js
@@ -20,29 +20,24 @@ import {
 const EDGES_ITERATION = [
   {
     name: 'edges',
-    counter: 'countEdges',
     type: 'mixed'
   },
   {
     name: 'inEdges',
-    counter: 'countInEdges',
     type: 'directed',
     direction: 'in'
   },
   {
     name: 'outEdges',
-    counter: 'countOutEdges',
     type: 'directed',
     direction: 'out'
   },
   {
     name: 'directedEdges',
-    counter: 'countDirectedEdges',
     type: 'directed'
   },
   {
     name: 'undirectedEdges',
-    counter: 'countUndirectedEdges',
     type: 'undirected'
   }
 ];
@@ -81,51 +76,6 @@ function collect(edges, object, key) {
 }
 
 /**
- * Function counting edges from the given object.
- *
- * @param  {object|undefined} object - Target object.
- * @param  {mixed}            [key]  - Optional key.
- * @return {number}                  - The number of found edges.
- */
-function count(object, key) {
-  let nb = 0;
-
-  const hasKey = arguments.length > 1;
-
-  if (!object || (hasKey && !(key in object)))
-    return nb;
-
-  if (hasKey)
-    return object[key] instanceof Set ? object[key].size : +!!object[key];
-
-  for (const k in object)
-    nb += (object[k] instanceof Set ? object[k].size : +!!object[k]);
-
-  return nb;
-}
-
-/**
- * Function used to count the edges of the given type.
- *
- * @param  {Graph}  graph - Target Graph instance.
- * @param  {string} type  - Type of edges to retrieve.
- * @return {number}
- */
-function countEdges(graph, type) {
-  if (type === 'mixed')
-    return graph.size;
-
-  let nb = 0;
-
-  graph._edges.forEach(data => {
-    if (!!data.undirected === (type === 'undirected'))
-      nb++;
-  });
-
-  return nb;
-}
-
-/**
  * Function creating an array of edge for the given type.
  *
  * @param  {Graph}   graph - Target Graph instance.
@@ -145,39 +95,6 @@ function createEdgeArray(graph, type) {
   });
 
   return list;
-}
-
-/**
- * Function counting the number of edges for the given type & the given node.
- *
- * @param  {Graph}   graph     - Target Graph instance.
- * @param  {string}  type      - Type of edges to retrieve.
- * @param  {string}  direction - In or out?
- * @param  {any}     node      - Target node.
- * @return {number}
- */
-function countEdgesForNode(graph, type, direction, node) {
-
-  // For this, we need to compute the "structure" index
-  graph.computeIndex('structure');
-
-  let nb = 0;
-
-  const nodeData = graph._nodes.get(node);
-
-  if (type !== 'undirected') {
-
-    if (direction !== 'out')
-      nb += count(nodeData.in);
-    if (direction !== 'in')
-      nb += count(nodeData.out);
-  }
-
-  if (type !== 'directed') {
-    nb += count(nodeData.undirected);
-  }
-
-  return nb;
 }
 
 /**
@@ -214,36 +131,6 @@ function createEdgeArrayForNode(graph, type, direction, node) {
 }
 
 /**
- * Function counting the number of edges for the given path.
- *
- * @param  {Graph}   graph  - Target Graph instance.
- * @param  {string}  type   - Type of edges to retrieve.
- * @param  {any}     source - Source node.
- * @param  {any}     target - Target node.
- * @return {array}          - Array of edges.
- */
-function countEdgesForPath(graph, type, source, target) {
-
-  // For this, we need to compute the "structure" index
-  graph.computeIndex('structure');
-
-  let nb = 0;
-
-  const sourceData = graph._nodes.get(source);
-
-  if (type !== 'undirected') {
-    nb += count(sourceData.in, target);
-    nb += count(sourceData.out, target);
-  }
-
-  if (type !== 'directed') {
-    nb += count(sourceData.undirected, target);
-  }
-
-  return nb;
-}
-
-/**
  * Function creating an array of edge for the given path.
  *
  * @param  {Graph}   graph  - Target Graph instance.
@@ -277,16 +164,14 @@ function createEdgeArrayForPath(graph, type, source, target) {
  * Function attaching an edge array creator method to the Graph prototype.
  *
  * @param {function} Class       - Target class.
- * @param {boolean}  counter     - Should we count or collect?
  * @param {object}   description - Method description.
  */
-function attachEdgeArrayCreator(Class, counter, description) {
+function attachEdgeArrayCreator(Class, description) {
   const {
+    name,
     type,
     direction
   } = description;
-
-  const name = counter ? description.counter : description.name;
 
   /**
    * Function returning an array or the count of certain edges.
@@ -314,9 +199,7 @@ function attachEdgeArrayCreator(Class, counter, description) {
       return [];
 
     if (!arguments.length)
-      return counter ?
-        countEdges(this, type) :
-        createEdgeArray(this, type);
+      return createEdgeArray(this, type);
 
     if (arguments.length === 1) {
 
@@ -324,9 +207,7 @@ function attachEdgeArrayCreator(Class, counter, description) {
         throw new NotFoundGraphError(`Graph.${name}: could not find the "${source}" node in the graph.`);
 
       // Iterating over a node's edges
-      return counter ?
-        countEdgesForNode(this, type, direction, source) :
-        createEdgeArrayForNode(this, type, direction, source);
+      return createEdgeArrayForNode(this, type, direction, source);
     }
 
     if (arguments.length === 2) {
@@ -346,11 +227,9 @@ function attachEdgeArrayCreator(Class, counter, description) {
 
       // If no such edge exist, we'll stop right there.
       if (!hasEdge)
-        return counter ? 0 : [];
+        return [];
 
-      return counter ?
-        countEdgesForPath(this, type, source, target) :
-        createEdgeArrayForPath(this, type, source, target);
+      return createEdgeArrayForPath(this, type, source, target);
     }
 
     throw new InvalidArgumentsGraphError(`Graph.${name}: too many arguments (expecting 0, 1 or 2 and got ${arguments.length}).`);
@@ -364,7 +243,6 @@ function attachEdgeArrayCreator(Class, counter, description) {
  */
 export function attachEdgeIterationMethods(Graph) {
   EDGES_ITERATION.forEach(description => {
-    attachEdgeArrayCreator(Graph, false, description);
-    attachEdgeArrayCreator(Graph, true, description);
+    attachEdgeArrayCreator(Graph, description);
   });
 }